% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_schema_files.R
\name{prepare_schema_files}
\alias{prepare_schema_files}
\title{Copy parameter files to directory}
\usage{
prepare_schema_files(directory)
}
\arguments{
\item{directory}{Directory to copy parameter files to.}
}
\value{
directory
}
\description{
Copy schema files from dependency packages to a (temporary) directory. This
is required, as \code{jsonvalidate} requires referenced schema files to exist in
the same directory as the referencing schema.
}
