% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_webapp_workflow.R
\name{run_webapp_workflow}
\alias{run_webapp_workflow}
\title{Run Webapp workflow}
\usage{
run_webapp_workflow(
  params,
  raw_params,
  run_analysis = TRUE,
  run_reporting = TRUE
)
}
\arguments{
\item{params}{(\code{list}) A list of parameters to be used in the analysis and
reporting process. See JSON Schema for details.}

\item{raw_params}{(\code{character}) Raw JSON string of parameters.}

\item{run_analysis}{(\code{logical}) Run the analysis process.}

\item{run_reporting}{(\code{logical}) Run the reporting process.}
}
\value{
returndes
}
\description{
Reun steps required to prepare a PACTA analysis and report.
}
